requirements.txt
Flask==3.0.0
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
openai==1.6.1
python-dotenv==1.0.0
gunicorn==21.2.0
